Make sure migrations work from fresh db:migrate

+++ db/migrate/20160509173152_remove_teams_and_team_memberships.rb
@@ -2,7 +2,7 @@ class RemoveTeamsAndTeamMemberships < ActiveRecord::Migration
   def change
     remove_reference :bots, :team
 
-    drop_table :team_memberships
-    drop_table :teams
+    drop_table :team_memberships if ActiveRecord::Base.connection.table_exists? 'team_memberships'
+    drop_table :teams if ActiveRecord::Base.connection.table_exists? 'teams'
   end
 end
